+++ website/src/pages/mod.rs
@@ -1,5 +1,7 @@
+mod guides_page;
 mod home_page;
 mod not_found_page;
 
+pub use guides_page::GuidesPage;
 pub use home_page::HomePage;
 pub use not_found_page::NotFoundPage;